refactor: pageStore을 entity로 이동 및 editorStore로 분리
