+++ apps/frontend/src/entities/page/model/pageStore.ts
@@ -1,0 +1,11 @@
+import { create } from "zustand";
+
+interface PageStore {
+  currentPage: number | null;
+  setCurrentPage: (currentPage: number | null) => void;
+}
+
+export const usePageStore = create<PageStore>((set) => ({
+  currentPage: null,
+  setCurrentPage: (currentPage: number | null) => set({ currentPage }),
+}));
