+++ apps/frontend/src/features/editor/index.ts
@@ -1,5 +1,8 @@
 export { onUpload } from "./api/imageUpload";
 
+export { useEditorStore } from "./model/editorStore";
+
+// TODO: 정리
 export { Editor } from "@/features/editor/ui/Editor";
 export { EditorTitle } from "@/features/editor/ui/EditorTitle";
 export { EditorActionPanel } from "@/features/editor/ui/EditorActionPanel";
