+++ apps/frontend/src/features/editor/model/editorStore.ts
@@ -1,27 +1,16 @@
 import { create } from "zustand";
 
-interface PageStore {
-  currentPage: number | null;
+interface EditorStore {
   isPanelOpen: boolean;
   isMaximized: boolean;
-  setCurrentPage: (currentPage: number | null) => void;
   togglePanel: () => void;
   toggleMaximized: () => void;
   setIsPanelOpen: (isOpen: boolean) => void;
 }
 
-export const usePageStore = create<PageStore>((set) => ({
-  currentPage: null,
+export const useEditorStore = create<EditorStore>((set) => ({
   isPanelOpen: true,
   isMaximized: false,
-  setCurrentPage: (currentPage: number | null) =>
-    set((state) => ({
-      currentPage,
-      isPanelOpen:
-        currentPage === state.currentPage
-          ? !state.isPanelOpen
-          : currentPage !== null,
-    })),
   togglePanel: () => set((state) => ({ isPanelOpen: !state.isPanelOpen })),
   toggleMaximized: () => set((state) => ({ isMaximized: !state.isMaximized })),
   setIsPanelOpen: (isPanelOpen: boolean) => set({ isPanelOpen }),
