+++ apps/frontend/src/features/pageSidebar/index.ts
@@ -1,7 +1,5 @@
 export { usePage, useCreatePage } from "./api/usePages";
 
-export { usePageStore } from "./model/pageStore";
-
 export { LogoBtn } from "./ui/LogoBtn";
 export { NoteList } from "./ui/NoteList";
 export { Tools } from "./ui/Tools";
